% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_manually.R
\name{cluster_manually}
\alias{cluster_manually}
\title{Cluster the cells using a manual cutoff}
\usage{
cluster_manually(ts, cutoff, transform = identity)
}
\arguments{
\item{ts}{A tissue slide object.}

\item{cutoff}{A value to use as the cutoff point for the manual classification of the cells.}

\item{transform}{A callable (function) the apply to the data before classification.}
}
\value{
The tissue slide object with a value in the "manual_classification" attribute.
}
\description{
Cluster the cells using a manual cutoff
}
