% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_slide.R
\name{get_manual_classification}
\alias{get_manual_classification}
\title{Get the manual classification from a tissue slide}
\usage{
get_manual_classification(ts)
}
\arguments{
\item{ts}{A tissue slide that has been manually classified.}
}
\value{
A \code{manual_classification} object.
}
\description{
Extract the manual classification from a tissue slide. This property should
not be set manually (ironically). Instead using the \code{cluster_manually()}
function. See the example below for a demonstration.
}
\examples{
pancreas_slide <- tissue_slide(pancreas)
pancreas_slide <- cluster_manually(pancreas_slide,
  fitc,
  cutoff = 4,
  transform = log10
)
get_manual_classification(pancreas_slide)
}
