% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_classification.R
\name{get_manual_classification_cutoff}
\alias{get_manual_classification_cutoff}
\title{Get the cutoff of a manual classification object.}
\usage{
get_manual_classification_cutoff(mc)
}
\arguments{
\item{mc}{A manual classification object.}
}
\value{
The cutoff value.
}
\description{
Get the cutoff of a manual classification object.
}
\examples{
mc <- manual_classification(
  c("a", "a", "b", "a", "b"),
  cutoff = 2.5,
  transform = log10
)
get_manual_classification_cutoff(mc)
}
