% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_slide.R
\name{get_slide_metadata}
\alias{get_slide_metadata}
\title{Get metadata from a tissue slide.}
\usage{
get_slide_metadata(ts)
}
\arguments{
\item{ts}{A Tissue slide object.}
}
\value{
The metadata object associated with the slide.
}
\description{
Extract the metadata stored with a tissue slide object. To set the metadata, use the function \code{set_slide_metadata()}.
}
\examples{
slide <- tissue_slide(pancreas, metadata = list(tissue = "pancreas", mouse = "OP24"))
get_slide_metadata(slide)
}
