% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancreas.R
\docType{data}
\name{pancreas}
\alias{pancreas}
\title{An example dataset from TUNEL on a pancreas.}
\format{
a tibble (67031 x 12)
\describe{
\item{cell_id}{a unique ID for each cell.}
\item{x}{x location}
\item{y}{y location}
\item{dapi}{DAPI channel intensity}
\item{fitc}{FITC channel intensity}
\item{area}{area of cell}
\item{major_axis_length}{length of the cell's longest dimension}
\item{minor_axis_length}{length of the cell's shortest dimension}
\item{eccentricity}{how circular the cell is}
\item{solidity}{how solid the cell is}
\item{extent}{unknown}
\item{orientation}{angle of the cell}
}
}
\usage{
pancreas
}
\description{
The output of running TUNEL on a pancreas slide.
}
\keyword{datasets}
