% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot the distribution of FITC intensities}
\usage{
plot_density(ts, value)
}
\arguments{
\item{ts}{Tissue slide object to plot.}

\item{value}{Values to plot the distribution of. (enquoted)}
}
\value{
A ggplot2 density plot.
}
\description{
Plot the distribution of FITC intensities
}
\examples{
pancreas_slide <- tissue_slide(pancreas)
plot_density(pancreas_slide, fitc)

}
