% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitc_density.R
\name{plot_fitc_density}
\alias{plot_fitc_density}
\title{Plot the distribution of FITC intensities}
\usage{
plot_fitc_density(ts)
}
\arguments{
\item{ts}{Tissue slide object to plot.}
}
\description{
Plot the distribution of FITC intensities
}
