% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slide_clusters.R
\name{plot_slide_clusters}
\alias{plot_slide_clusters}
\title{Plot the clusters on the slide}
\usage{
plot_slide_clusters(ts, method = "manual", prop = 0.25)
}
\arguments{
\item{ts}{A tissue slide object}

\item{method}{The classification method to use. The classification should have already been done.}

\item{prop}{The proportion of cells to include in the plot (too many cells can impact performance).}
}
\description{
Plot the clusters on the slide
}
