% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tissue.R
\name{plot_tissue}
\alias{plot_tissue}
\title{Plot the tissue slide showing the FITC intensity}
\usage{
plot_tissue(ts, prop = 0.25)
}
\arguments{
\item{ts}{Tissue slide object to plot.}

\item{prop}{The proportion of cells to plot (too many cells can slow performance).}
}
\description{
Plots the slide with each point representing a cell and the color represents the intensity of FITC.
}
