% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_slide.R
\name{set_slide_metadata}
\alias{set_slide_metadata}
\title{Set metadata from a tissue slide.}
\usage{
set_slide_metadata(ts, metadata)
}
\arguments{
\item{ts}{A Tissue slide object.}

\item{metadata}{Metadata of arbitrary structure.}
}
\value{
Returns the tissue slide object with the metadata.
}
\description{
Write the metadata stored with a tissue slide object. To get the metadata, use the function \code{get_slide_metadata()}.
}
\examples{
slide <- tissue_slide(pancreas, metadata = list(tissue = "pancreas", mouse = "OP24"))
slide <- set_slide_metadata(slide, list(tissue = "not pancreas", mouse = "Mickey"))
get_slide_metadata(slide)
}
