% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tissue_plot.R
\name{standard_tissue_plot}
\alias{standard_tissue_plot}
\title{Some standard modifications to tissue slide plots. (not exported)}
\usage{
standard_tissue_plot(p)
}
\arguments{
\item{p}{Original plot}
}
\value{
A modified ggplot2 object
}
\description{
Some standard modifications to tissue slide plots. (not exported)
}
