% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_cluster_results.R
\name{summarize_cluster_results}
\alias{summarize_cluster_results}
\title{Summarize the cluster results}
\usage{
summarize_cluster_results(ts)
}
\arguments{
\item{ts}{The tissue slide}
}
\value{
A data.frame of the summarizesed results, one column per classification method.
}
\description{
Create a tibble that summarizes the classification of cells.
}
