% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tissue_slide.R
\name{theme_tissue_slide}
\alias{theme_tissue_slide}
\title{A ggplot2 theme for tissue slide plots}
\usage{
theme_tissue_slide(base_size = 11, base_family = "Helvetica")
}
\arguments{
\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme.
}
\description{
A ggplot2 theme for tissue slide plots
}
\examples{
ts <- tissue_slide(pancreas)
plot_tissue(ts) + theme_tissue_slide()
}
