% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_slide.R
\name{tissue_slide}
\alias{tissue_slide}
\title{Tissue slide information}
\usage{
tissue_slide(slide_data)
}
\arguments{
\item{slide_data}{The data from the slide.}
}
\description{
An S3 class to hold the information for an individual slide.
}
