% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_classification.R
\name{manual_classification}
\alias{manual_classification}
\title{Manual classification object.}
\usage{
manual_classification(
  assignments = factor(),
  cutoff = NA_real_,
  transform = identity
)
}
\arguments{
\item{assignments}{A factor vector of cluster assignments.}

\item{cutoff}{The value used for the manual classification.}

\item{transform}{A callable (function) to transform the data before applying the cutoff classification.}
}
\value{
A manual classification that behaves like a factor.
}
\description{
An S3 class for handling manual classification.
}
\examples{
# An empty classification.
manual_classification()

# The numeric assignments will be turned into a factor.
manual_classification(c(1, 2, 1, 2, 1, 1), cutoff = 4000)

# Provide a different function used to transform the data before
# applying the cutoff.
manual_classification(
  c("a", "a", "b", "a", "b"),
  cutoff = 2.5,
  transform = log10
)
}
