% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slide_clusters.R
\name{plot_slide_clusters}
\alias{plot_slide_clusters}
\title{Plot the clusters on the slide}
\usage{
plot_slide_clusters(
  ts,
  method = "manual",
  prop = 0.25,
  point_size = 0.5,
  alpha_values = c(0.2, 0.7)
)
}
\arguments{
\item{ts}{A tissue slide object}

\item{method}{The classification method to use. The classification should have already been done.}

\item{prop}{The proportion of cells to include in the plot (too many cells can impact performance).}
}
\value{
A ggplot2 object.
}
\description{
Plot the clusters on the slide
}
\examples{
pancreas_slide <- tissue_slide(pancreas)
pancreas_slide <- cluster_manually(pancreas_slide,
  fitc,
  cutoff = 4,
  transform = log10
)
plot_slide_clusters(pancreas_slide)
}
