% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_slide.R
\name{tissue_slide}
\alias{tissue_slide}
\title{Tissue slide information}
\usage{
tissue_slide(slide_data, metadata = list())
}
\arguments{
\item{slide_data}{The data from the slide.}

\item{metadata}{Additional metadata as a list with arbitrary structure and values.}
}
\value{
A \code{tissue_slide} object that is a subclass of a tibble.
}
\description{
An S3 class to hold the information for an individual slide.
}
\examples{
pancreas_slide <- tissue_slide(pancreas)
}
